Added End Effector Design V10
This new end effector supports the LDO Orbiter V2.x direct extruder. In combination with the Orbiter Toolboard this allows for a complete remote toolhead setup. There are two versions of the end effector, one with two neopixels and one without. For this new end effector, you will also need a new electronics compartment for the Orbiter Toolboard's USB Adapter and the Orbiter USB/Power cable.